+++ compiler/src/test/scala/dx/compiler/CompilerTest.scala
@@ -127,6 +127,22 @@ class CompilerTest extends AnyFlatSpec with Matchers with BeforeAndAfterAll {
     }
   }
 
+  it should "compile a task with dynamic instance type selection" in {
+    def compile(instanceTypeSelection: String): Map[String, JsValue] = {
+      val path = pathFromBasename("compiler", "add.wdl")
+      val args = path.toString :: "-instanceTypeSelection" :: instanceTypeSelection :: cFlags
+      val retval = Main.compile(args.toVector)
+      val appletId = retval match {
+        case SuccessfulCompileNativeNoTree(_, Vector(appletId)) => appletId
+        case other =>
+          throw new Exception(s"expected successful compilation of a single applet not ${other}")
+      }
+      dxApi.applet(appletId).describe(Set(Field.Details)).details.get.asJsObject.fields
+    }
+    compile("static") should contain key Constants.InstanceTypeDb
+    compile("dynamic") shouldNot contain key Constants.InstanceTypeDb
+  }
+
   it should "Native compile a linear WDL workflow" taggedAs NativeTest in {
     val path = pathFromBasename("compiler", "wf_linear.wdl")
     val args = path.toString :: cFlags